+++ src/demo/views/Components.vue
@@ -13,7 +13,9 @@
     <Tag class="mt-10" />
     <TagGroup class="mt-10" />
     <Avatar class="mt-10" />
+    <Tooltip class="mt-10" />
     <Selects class="mt-10" />
+    <TextAreas class="mt-10" />
   </div>
 </template>
 
@@ -31,7 +33,9 @@ import TagGroup from '../sections/TagGroup.vue'
 import Tag from '../sections/Tag.vue'
 import Popups from '../sections/Popups.vue'
 import Sliders from '../sections/Sliders.vue'
+import Tooltip from '../sections/Tooltip.vue'
 import Selects from '../sections/Selects.vue'
+import TextAreas from '../sections/TextAreas.vue'
 
 @Options({
   components: {
@@ -47,8 +51,9 @@ import Selects from '../sections/Selects.vue'
     Selects,
     TabSection,
     Tag,
-    TagGroup
-
+    TagGroup,
+    TextAreas,
+    Tooltip
   }
 })
 export default class App extends Vue {}